moveonly: add mp/type-function.h
